myntist-beacon/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
boto3==1.34.95
botocore==1.34.95
requests==2.31.0
PyGithub==2.3.0
PyNaCl==1.5.0
prometheus-client==0.20.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-httpx==0.30.0
python-dotenv==1.0.1
feedgen==1.0.0
cryptography==42.0.7
base58==2.1.1
PyYAML==6.0.1
pyarrow==16.0.0
